+++ server/src/main/java/org/opensearch/index/engine/EngineConfig.java
@@ -82,6 +82,7 @@ public final class EngineConfig {
     private volatile boolean enableGcDeletes = true;
     private final TimeValue flushMergesAfter;
     private final String codecName;
+    private final int compressionLevel;
     private final ThreadPool threadPool;
     private final Engine.Warmer warmer;
     private final Store store;
@@ -142,6 +143,13 @@ public final class EngineConfig {
                 return s;
         }
     }, Property.IndexScope, Property.NodeScope);
+    public static final Setting<Integer> INDEX_CODEC_COMPRESSION_LEVEL_SETTING = Setting.intSetting(
+        "index.codec.compression_level",
+        6,
+        1,
+        6,
+        Property.IndexScope
+    );
 
     /**
      * Configures an index to optimize documents with auto generated ids for append only. If this setting is updated from <code>false</code>
@@ -179,6 +187,7 @@ public final class EngineConfig {
         this.codecService = builder.codecService;
         this.eventListener = builder.eventListener;
         codecName = builder.indexSettings.getValue(INDEX_CODEC_SETTING);
+        compressionLevel = builder.indexSettings.getValue(INDEX_CODEC_COMPRESSION_LEVEL_SETTING);
         // We need to make the indexing buffer for this shard at least as large
         // as the amount of memory that is available for all engines on the
         // local node so that decisions to flush segments to disk are made by
@@ -250,6 +259,9 @@ public final class EngineConfig {
      * </p>
      */
     public Codec getCodec() {
+        if (codecName.equals(CodecService.ZSTD_CODEC) || codecName.equals(CodecService.ZSTD_NO_DICT_CODEC)) {
+            return codecService.codec(codecName, compressionLevel);
+        }
         return codecService.codec(codecName);
     }